veloce-backend/pyproject.toml
[project]
name = "veloce-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=1.27.0,<2.0.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "motor>=3.6.0",
    "openai>=1.57.0",
    "bcrypt>=4.0.1",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.17",
    "uvicorn[standard]>=0.32.0",
]
